+++ RELEASE-NOTES.md
@@ -1,5 +1,6 @@
 # Release Notes for protobuf-java-format
 * 1.5
+   * [#35](https://github.com/bivas/protobuf-java-format/issues/35) - Fix character encoding bugs when printing to strings and byte arrays.
    * [#37](https://github.com/bivas/protobuf-java-format/pull/37) - Fix NPE during json deserialization.
 * 1.4
    * [#15](https://github.com/bivas/protobuf-java-format/issues/15) - Fix unsigned values for JacksonFormatter.